fix(breakout-room): isolate timer state to dedicated memo components
