+++ bigbluebutton-html5/imports/ui/components/breakout-room/breakout-room/breakout-timer-editor/component.tsx
@@ -1,0 +1,186 @@
+import React, {
+  useState, useCallback, useRef, useEffect,
+} from 'react';
+import { defineMessages, useIntl } from 'react-intl';
+import { useMutation } from '@apollo/client';
+import useTimeSync from '/imports/ui/core/local-states/useTimeSync';
+import Styled from '../running-room/styles';
+import { BREAKOUT_ROOM_SET_TIME } from '../../mutations';
+
+const DEBOUNCE_DELAY = 1500;
+
+const intlMessages = defineMessages({
+  durationOfBreakout: {
+    id: 'app.createBreakoutRoom.durationOfBreakout',
+    description: 'Duration of Breakout Rooms label',
+  },
+});
+
+interface BreakoutTimerEditorProps {
+  breakoutDurationInSeconds: number;
+  breakoutStartedAt: number;
+}
+
+const BreakoutTimerEditorComponent: React.FC<BreakoutTimerEditorProps> = ({
+  breakoutDurationInSeconds, breakoutStartedAt,
+}) => {
+  const intl = useIntl();
+  const [timeSync] = useTimeSync();
+  const timerIntervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
+  const debounceTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+  const innerTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+  const [selectedTimerField, setSelectedTimerField] = useState<'h' | 'm' | 's'>('m');
+  const [remainingTime, setRemainingTime] = useState<number>(0);
+  const [editingTime, setEditingTime] = useState<number | null>(null);
+  const [breakoutRoomSetTime] = useMutation(BREAKOUT_ROOM_SET_TIME);
+
+  useEffect(() => {
+    const calcRemaining = () => {
+      if (!Number.isFinite(breakoutStartedAt) || breakoutStartedAt === 0
+          || !Number.isFinite(breakoutDurationInSeconds)) {
+        return 0;
+      }
+      const now = Date.now() + timeSync;
+      const end = breakoutStartedAt + (breakoutDurationInSeconds * 1000);
+      return Math.max(0, Math.floor((end - now) / 1000));
+    };
+
+    if (editingTime === null) {
+      setRemainingTime(calcRemaining());
+    }
+
+    timerIntervalRef.current = setInterval(() => {
+      if (editingTime === null) {
+        setRemainingTime(calcRemaining());
+      }
+    }, 1000);
+
+    return () => {
+      if (timerIntervalRef.current) clearInterval(timerIntervalRef.current);
+    };
+  }, [breakoutDurationInSeconds, breakoutStartedAt, timeSync, editingTime]);
+
+  useEffect(() => () => {
+    if (debounceTimerRef.current) clearTimeout(debounceTimerRef.current);
+    if (innerTimerRef.current) clearTimeout(innerTimerRef.current);
+  }, []);
+
+  const displayedTime = editingTime ?? remainingTime;
+  const hours = Math.floor(displayedTime / 3600);
+  const minutes = Math.floor((displayedTime % 3600) / 60);
+  const seconds = displayedTime % 60;
+  const padNum = (n: number) => n.toString().padStart(2, '0');
+
+  const commitTimeChange = useCallback((newTotalSeconds: number) => {
+    const clamped = Math.max(60, newTotalSeconds);
+    setEditingTime(clamped);
+
+    if (debounceTimerRef.current) clearTimeout(debounceTimerRef.current);
+
+    debounceTimerRef.current = setTimeout(() => {
+      const newMinutes = Math.max(1, Math.ceil(clamped / 60));
+      breakoutRoomSetTime({ variables: { timeInMinutes: newMinutes } });
+      if (innerTimerRef.current) clearTimeout(innerTimerRef.current);
+      innerTimerRef.current = setTimeout(() => setEditingTime(null), 500);
+    }, DEBOUNCE_DELAY);
+  }, [breakoutRoomSetTime]);
+
+  const adjustTime = useCallback((delta: number) => {
+    let deltaSeconds = 0;
+    if (selectedTimerField === 'h') deltaSeconds = delta * 3600;
+    else if (selectedTimerField === 'm') deltaSeconds = delta * 60;
+    else deltaSeconds = delta;
+
+    const base = editingTime ?? remainingTime;
+    commitTimeChange(base + deltaSeconds);
+  }, [selectedTimerField, editingTime, remainingTime, commitTimeChange]);
+
+  const handleTimerInputChange = useCallback((
+    field: 'h' | 'm' | 's',
+    value: number,
+  ) => {
+    let newH = hours;
+    let newM = minutes;
+    let newS = seconds;
+
+    if (field === 'h') newH = Math.max(0, Math.min(23, value));
+    else if (field === 'm') newM = Math.max(0, Math.min(59, value));
+    else newS = Math.max(0, Math.min(59, value));
+
+    const totalSeconds = (newH * 3600) + (newM * 60) + newS;
+    commitTimeChange(totalSeconds);
+  }, [hours, minutes, seconds, commitTimeChange]);
+
+  return (
+    <Styled.TimerSection>
+      <Styled.TimerLabel>
+        {intl.formatMessage(intlMessages.durationOfBreakout)}
+      </Styled.TimerLabel>
+      <Styled.TimerRow>
+        <Styled.TimerTimeBtn
+          $variant="minus"
+          onClick={() => adjustTime(-1)}
+          aria-label="Decrease time"
+          data-test="decreaseBreakoutTimeButton"
+        >
+          −
+        </Styled.TimerTimeBtn>
+        <Styled.TimerDisplay data-test="breakoutRemainingTime">
+          <Styled.TimerInput
+            type="number"
+            min={0}
+            max={23}
+            value={padNum(hours)}
+            $selected={selectedTimerField === 'h'}
+            onClick={() => setSelectedTimerField('h')}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              handleTimerInputChange('h', Number(e.target.value));
+            }}
+            aria-label="Hours"
+            data-test="breakoutRoomTimerHoursInput"
+          />
+          <Styled.TimerColon>:</Styled.TimerColon>
+          <Styled.TimerInput
+            type="number"
+            min={0}
+            max={59}
+            value={padNum(minutes)}
+            $selected={selectedTimerField === 'm'}
+            onClick={() => setSelectedTimerField('m')}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              handleTimerInputChange('m', Number(e.target.value));
+            }}
+            aria-label="Minutes"
+            data-test="breakoutRoomTimerMinutesInput"
+          />
+          <Styled.TimerColon>:</Styled.TimerColon>
+          <Styled.TimerInput
+            type="number"
+            min={0}
+            max={59}
+            value={padNum(seconds)}
+            $selected={selectedTimerField === 's'}
+            onClick={() => setSelectedTimerField('s')}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              handleTimerInputChange('s', Number(e.target.value));
+            }}
+            aria-label="Seconds"
+          />
+        </Styled.TimerDisplay>
+        <Styled.TimerTimeBtn
+          $variant="plus"
+          onClick={() => adjustTime(1)}
+          aria-label="Increase time"
+          data-test="increaseBreakoutTimeButton"
+        >
+          +
+        </Styled.TimerTimeBtn>
+      </Styled.TimerRow>
+    </Styled.TimerSection>
+  );
+};
+
+const BreakoutTimerEditor = React.memo(BreakoutTimerEditorComponent);
+BreakoutTimerEditor.displayName = 'BreakoutTimerEditor';
+
+export default BreakoutTimerEditor;
